% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approShapley.R
\name{approShapley}
\alias{approShapley}
\title{Monte Carlo Simulation based on the 2009 paper by Javier Castro}
\usage{
approShapley(n, m, v)
}
\arguments{
\item{n}{The number of players}

\item{m}{The number of samples}

\item{v}{characteristic function}
}
\value{
List of Shapley values
}
\description{
Monte Carlo Simulation based on the 2009 paper by Javier Castro
}
\examples{
print (ApproShapley (10, 10000, gloveGame(1:5,6:10)))
}
