% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAirportCostVector.R
\name{buildAirportCostVector}
\alias{buildAirportCostVector}
\title{Returns the cost vector for an airport game}
\usage{
buildAirportCostVector(cost_info)
}
\arguments{
\item{cost_info}{A list of tuples that describes how often each number is going to appear ((number, times), ...). The tuples in cost_info should be sorted by the cost in ascending order.}
}
\value{
The cost vector described by cost_info.
}
\description{
Returns the cost vector for an airport game
}
\examples{
costs <- buildAirportCostVector(list(c(1, 8), c(2, 12), c(3, 6), c(4, 14), c(5, 8), c(6, 9), c(7, 13), c(8, 10), c(9, 10), c(10, 10)))
}
