% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceBanzhaff.R
\name{confidenceBanzhaff}
\alias{confidenceBanzhaff}
\title{Monte Carlo Simulation for calculating the Banzhaff value in a simple game based on the paper "Approximating power indices: theoretical and empirical analysis" by Bachrach et al. from 2010}
\usage{
confidenceBanzhaff(i, n, v, conf, w)
}
\arguments{
\item{i}{Player}

\item{n}{Number of players}

\item{v}{Characteristic function}

\item{conf}{Confidence of the true Banzhaff value being in the confidence interval}

\item{w}{Width of the confidence interval}
}
\value{
Confidence interval as a vector of size 2
}
\description{
Monte Carlo Simulation for calculating the Banzhaff value in a simple game based on the paper "Approximating power indices: theoretical and empirical analysis" by Bachrach et al. from 2010
}
\examples{
confidenceBanzhaff(3, 4, weightedVotingGame(c(1, 1, 2, 3), 1/2), 0.95, 0.01)
}
