% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromICoalitionToCoalition.R
\name{fromICoalitionToCoalition}
\alias{fromICoalitionToCoalition}
\title{Access possible coalitions of players through index; based on paper: "Statistics and game theory: Estimating coalitional values in R software" (Alejandro Saavedra-Nieves, 2020)}
\usage{
fromICoalitionToCoalition(n, iCoalition)
}
\arguments{
\item{n}{indicates the cardinality of the set of players N}

\item{iCoalition}{indicates the index of the coalition. This value should be greater than or equal to 1 and less than or equal to 2^n}
}
\value{
numeric binary vector, of dimension n, which each component i indicates the belonging of player i to the coalition
}
\description{
Access possible coalitions of players through index; based on paper: "Statistics and game theory: Estimating coalitional values in R software" (Alejandro Saavedra-Nieves, 2020)
}
\examples{
print(fromICoalitionToCoalition(5, 27))
}
