% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stApproOwenAndBanzhafOwen.R
\name{stApproOwenAndBanzhafOwen}
\alias{stApproOwenAndBanzhafOwen}
\title{A two-stage procedure to estimate the Banzhaf–Owen value based on paper: "On stratified sampling for estimating coalitional values" (A. Saavedra-Nieves, 2022)}
\usage{
stApproOwenAndBanzhafOwen(i, l, P, v)
}
\arguments{
\item{i}{indicates player}

\item{l}{set total sample size}

\item{P}{A priori unions}

\item{v}{TU-game with a system of a priori unions}
}
\value{
approximation of the Banzaf-Owen value and Owen value based on stratified sampling
}
\description{
A two-stage procedure to estimate the Banzhaf–Owen value based on paper: "On stratified sampling for estimating coalitional values" (A. Saavedra-Nieves, 2022)
}
\examples{
print (stApproOwenAndBanzhafOwen (1, 10000, list(c(1,2),(3)), gloveGame(1:2,3:3)))
print (stApproOwenAndBanzhafOwen (3, 10000, list(c(1), c(2, 3)), gloveGame(1:2,3:3))$Owen)
}
