% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoStageApproBanzhafOwen.R
\name{twoStageApproBanzhafOwen}
\alias{twoStageApproBanzhafOwen}
\title{A two-stage procedure to estimate the Banzhaf–Owen value}
\usage{
twoStageApproBanzhafOwen(i, lr, ls, P, v)
}
\arguments{
\item{i}{indicates player}

\item{lr}{number of samples of coalitions of unions}

\item{ls}{number of samples of coalitions in Pi for each element of sample}

\item{P}{A priori unions}

\item{v}{TU-game with a system of a priori unions}
}
\value{
approximation of the Banzaf-Owen value based on two-stage sampling
}
\description{
twoStageApproBanzhafOwen is a sampling method to approximate the Banzhaf-Owen value for
a specified TU game with a system of a priori unions, based on two-stage sampling.
Based on: "Sampling methods to estimate the Banzhaf–Owen value" (A. Saavedra-Nieves & M. G. Fiestras-Janeiro, 2020)
}
\examples{
print(twoStageApproBanzhafOwen(1, 2, 2, list(c(1, 2), (3)), gloveGame(1:2, 3:3)))

}
